requirement.txt
python == 3.8.5
opencv2
Numpy
